separated models from actions
