+++ actions/plugins.py
@@ -1,11 +1,8 @@
 from typing import List, Any
-from pydantic import BaseModel
 from . import action_store as action_store
 from .http_wrapper import get_wrapper
+from models.plugin_models import Plugin
 
-class Plugin(BaseModel):
-    shortName: str
-    longName: str
 
 @action_store.kubiya_action()
 def list_jenkins_plugins(_: Any = None) -> List[Plugin]:
