+++ models/node_models.py
@@ -1,0 +1,5 @@
+from pydantic import BaseModel
+
+class JenkinsNode(BaseModel):
+    displayName: str
+    offline: bool
