+++ models/pipeline_models.py
@@ -1,0 +1,5 @@
+from pydantic import BaseModel
+
+class JenkinsPipeline(BaseModel):
+    name: str
+    url: str
