+++ models/plugin_models.py
@@ -1,0 +1,6 @@
+from pydantic import BaseModel
+
+
+class Plugin(BaseModel):
+    shortName: str
+    longName: str
